fastapi-agent-app/requirements.txt
fastapi
uvicorn
google-cloud-logging
python-dotenv